# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "webFinal API",
        "contact": {},
        "license": {
            "name": "GPL 3.0"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/msg/v1/delete/{tid}": {
            "get": {
                "description": "message api\nuser can delete exist msg with message id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "msg"
                ],
                "summary": "Delete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "target id",
                        "name": "tid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Msg"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Msg"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/msg/v1/get/{tid}": {
            "get": {
                "description": "message api\nuser can get exist msg with target id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "msg"
                ],
                "summary": "Get",
                "parameters": [
                    {
                        "type": "string",
                        "description": "target id",
                        "name": "tid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Msg"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Msg"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/api/v1/msg/v1/recall/{tid}": {
            "get": {
                "description": "message api\nuser can recall msg sent in 10 minutes",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "msg"
                ],
                "summary": "Recall",
                "parameters": [
                    {
                        "type": "string",
                        "description": "target id",
                        "name": "tid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Msg"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Msg"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/msg/v1/send/{tid}": {
            "post": {
                "description": "message api\nuser can send msg with name",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "msg"
                ],
                "summary": "Send",
                "parameters": [
                    {
                        "type": "string",
                        "description": "target id",
                        "name": "tid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Msg"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Msg"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "models.Msg": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "isRecall": {
                    "type": "boolean"
                },
                "receiveUserID": {
                    "type": "integer"
                },
                "receiveVisible": {
                    "type": "boolean"
                },
                "sendTime": {
                    "type": "integer"
                },
                "sendUserID": {
                    "type": "integer"
                },
                "sendVisible": {
                    "type": "boolean"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "web.Msg": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
    }
}